吴岳欣 初步解决登陆异步问题

+++ code/elife/news-server/src/main/java/newsserver/controller/NoticeController.java
@@ -33,66 +33,84 @@ public class NoticeController {
 
     @RequestMapping(path = "/addNotice")
     @ResponseBody
-    public JSONObject addNotice(@RequestParam String content, @RequestParam String managerName, @RequestParam int communityId, @RequestParam String username, @RequestParam int isMass){
-        net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
-        if("".equals(content)||"".equals(managerName)||communityId==0){
-            object.put("addNotice", "0");
-            object.put("message","信息不能为空");
+    public JSONObject addNotice(HttpServletRequest request, @RequestParam String content, @RequestParam String managerName, @RequestParam int communityId, @RequestParam String username, @RequestParam int isMass){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("1".equals(role)) || !name.equals(managerName)) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        } else {
+            net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
+            if ("".equals(content) || "".equals(managerName) || communityId == 0) {
+                object.put("addNotice", "0");
+                object.put("message", "信息不能为空");
+                return object;
+            }
+            SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+            String noticeTime = df.format(new Date());
+            Notice notice = new Notice(noticeTime, content, managerName, communityId);
+            /**先在notice数据库中存*/
+            int noticeId = noticeService.save(notice);
+
+            if (isMass == 0) {
+                NoticeUser noticeUser = new NoticeUser(noticeId, username);
+                noticeService.saveNoticeUser(noticeUser);
+
+            } else {
+                List<String> usernameList = noticeService.findUsernameByCommunityId(communityId);
+                int length = usernameList.size();
+                for (int i = 0; i < length; i++) {
+                    NoticeUser noticeUser1 = new NoticeUser(noticeId, usernameList.get(i));
+                    noticeService.saveNoticeUser(noticeUser1);
+                }
+            }
+            object.put("addNotice", "1");
+            object.put("message", "发布物业通知成功");
             return object;
         }
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        String noticeTime = df.format(new Date());
-        Notice notice=new Notice(noticeTime,content,managerName,communityId);
-        /**先在notice数据库中存*/
-         int noticeId = noticeService.save(notice);
-
-         if(isMass==0){
-             NoticeUser noticeUser=new NoticeUser(noticeId,username);
-             noticeService.saveNoticeUser(noticeUser);
-
-         }else {
-             List<String> usernameList=noticeService.findUsernameByCommunityId(communityId);
-             int length=usernameList.size();
-             for (int i=0;i<length;i++){
-                 NoticeUser noticeUser1=new NoticeUser(noticeId,usernameList.get(i));
-                 noticeService.saveNoticeUser(noticeUser1);
-             }
-         }
-        object.put("addNotice", "1");
-        object.put("message","发布物业通知成功");
-        return object;
     }
     @RequestMapping(path = "/ManagerFindNotice")
     @ResponseBody
-    public JSONArray  managerFindNotice(@RequestParam String managerName, @RequestParam int pageNumber, @RequestParam int pageSize){
-
-        List<Notice> notices=noticeService.managerFindNotice(managerName,pageNumber,pageSize);
-        int count = noticeService.countByManagerName(managerName);
-        int pageNum = count/pageSize;
-        if (count%pageSize!=0){
-            pageNum += 1;
-        }
-        JSONObject firstObject = new JSONObject();
-        firstObject.put("pageNum",pageNum);
-        JSONArray result = new JSONArray();
-        result.add(firstObject);
-        for(Notice notice : notices){
-            JSONObject object = new JSONObject();
-            object.put("id", notice.getNoticeId());
-            object.put("time", notice.getNoticeTime());
-            object.put("content",notice.getNoticeContent());
-            object.put("managename",notice.getManagerName());
-            object.put("communityId",notice.getCommunityId());
-            List<NoticeUser> noticeUserList=noticeService.findbyNoticeId(notice.getNoticeId());
-            if(noticeUserList.size()>1){
-                object.put("receiver","全部");
+    public JSONArray  managerFindNotice(HttpServletRequest request, @RequestParam String managerName, @RequestParam int pageNumber, @RequestParam int pageSize){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        JSONArray jsonArray = new JSONArray();
+        if (StringUtils.isEmpty(name) || !("1".equals(role)) || !(name.equals(managerName))) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            jsonArray.add(jsonObject);
+            return jsonArray;
+        } else {
+            List<Notice> notices = noticeService.managerFindNotice(managerName, pageNumber, pageSize);
+            int count = noticeService.countByManagerName(managerName);
+            int pageNum = count / pageSize;
+            if (count % pageSize != 0) {
+                pageNum += 1;
             }
-            else if(noticeUserList.size()>0){
-                object.put("receiver",noticeUserList.get(0).getUsername());
+            JSONObject firstObject = new JSONObject();
+            firstObject.put("pageNum", pageNum);
+            JSONArray result = new JSONArray();
+            result.add(firstObject);
+            for (Notice notice : notices) {
+                JSONObject object = new JSONObject();
+                object.put("id", notice.getNoticeId());
+                object.put("time", notice.getNoticeTime());
+                object.put("content", notice.getNoticeContent());
+                object.put("managename", notice.getManagerName());
+                object.put("communityId", notice.getCommunityId());
+                List<NoticeUser> noticeUserList = noticeService.findbyNoticeId(notice.getNoticeId());
+                if (noticeUserList.size() > 1) {
+                    object.put("receiver", "全部");
+                } else if (noticeUserList.size() > 0) {
+                    object.put("receiver", noticeUserList.get(0).getUsername());
+                }
+                result.appendElement(object);
             }
-            result.appendElement(object);
+            return result;
         }
-        return result;
     }
     @RequestMapping(path = "/findMyNotice")
     @ResponseBody
@@ -101,7 +119,7 @@ public class NoticeController {
         String name = (String) session.getAttribute("username");
         String role = (String) session.getAttribute("role");
         JSONArray jsonArray = new JSONArray();
-        if (StringUtils.isEmpty(name) || !("0".equals(role))) {
+        if (StringUtils.isEmpty(name) || !("0".equals(role))  || !(name.equals(username))) {
             JSONObject jsonObject = new JSONObject();
             jsonObject.put("login", 0);
             jsonArray.add(jsonObject);
@@ -123,18 +141,25 @@ public class NoticeController {
      * 管理员功能，删除数据库中一条物业信息，包括notice表中和noticeUser表中*/
     @RequestMapping(path = "/deleteOneNotice")
     @ResponseBody
-    public JSONObject deleteOneNotice(@RequestParam int noticeId){
-        net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
-        if(noticeService.deleteByNotcieId(noticeId).equals("删除这一条物业信息成功")){
-            object.put("deleteOneNotice", "1");
-            object.put("message","删除这一条物业信息成功");
+    public JSONObject deleteOneNotice(HttpServletRequest request, @RequestParam int noticeId){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("1".equals(role))) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        } else {
+            net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
+            if (noticeService.deleteByNotcieId(noticeId).equals("删除这一条物业信息成功")) {
+                object.put("deleteOneNotice", "1");
+                object.put("message", "删除这一条物业信息成功");
+                return object;
+            }
+            object.put("deleteOneNotice", "0");
+            object.put("message", "删除这一条物业信息失败");
             return object;
-
         }
-        object.put("deleteOneNotice", "0");
-        object.put("message","删除这一条物业信息失败");
-        return object;
-
     }
     /**
      * 用户功能，删除我的通知列表中的某一条通知*/
@@ -144,7 +169,7 @@ public class NoticeController {
         HttpSession session = request.getSession();
         String name = (String) session.getAttribute("username");
         String role = (String) session.getAttribute("role");
-        if (StringUtils.isEmpty(name) || !("0".equals(role))) {
+        if (StringUtils.isEmpty(name) || !("0".equals(role)) || !name.equals(username)) {
             JSONObject jsonObject = new JSONObject();
             jsonObject.put("login", 0);
             return jsonObject;
@@ -165,12 +190,20 @@ public class NoticeController {
      * 用户功能，删除我的通知列表中的所有条通知*/
     @RequestMapping(path = "/deleteMyNotice")
     @ResponseBody
-    public JSONObject deleteMyNotice(@RequestParam String username){
-        noticeService.deleteAllByUsername(username);
-        net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
-        object.put("deleteMyNotice", "1");
-        object.put("message","删除我的所有通知成功");
-        return object;
-
+    public JSONObject deleteMyNotice(HttpServletRequest request, @RequestParam String username){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("0".equals(role)) || !name.equals(username)) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        } else {
+            noticeService.deleteAllByUsername(username);
+            net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
+            object.put("deleteMyNotice", "1");
+            object.put("message", "删除我的所有通知成功");
+            return object;
+        }
     }
 }
